fix: add better routing system - Added home.js for main route of the game - Moved routes from index.js to home.js
